refactor(discord): consolidate organization commands into single OrganizationCommand
Consolidates create-organization, list-organizations, and join-organization
commands into a single /organization command with subcommands:
- /organization create <name> [slug] [webhook]: Create new organization
- /organization list: List all organizations
- /organization join <org>: Join an organization

Groups organization-related commands under single domain command.
